requirements.txt
streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
whoosh>=2.7.4
nltk>=3.8
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
